requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-extensions==3.1.5
django-import-export==2.6.1
django-ratelimit==3.0.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
gunicorn-color==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
python-dateutil==2.8.2
python-rapidjson==1.5
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
termcolor==1.1.0
xlrd==2.0.1
xlwt==1.3.0
